things are working quite nicely!
There are of course plenty of bugs and missing features. But most orders are now completed in a sensible way, so that's good! The next steps is to restructure datastructures etc. and then add proper handling of losing network etc. Will probably be quite some work before next week, but we think we'll manage ;D