the sun is starting to rise
(as in, we're nearing completion. Time to go home for the night). A lot of things are up and running! There are still some bugs though, and some things that's not yet implemented. Such as how to handle the elevator when it doesn't have network, and to make sure that everything is handled correctly when motor is disabled / similiar.